requirements.txt
openai>=1.14.0
python-telegram-bot>=21.0
APScheduler>=3.10.4,<3.12.0
httpx>=0.28
python-dotenv>=1.0.0
requests>=2.31.0
